Refactoring codebase to execute common flows
