+++ components/migration-client/wso2-api-migration-client/src/main/java/org/wso2/carbon/apimgt/migration/client/MigrateFrom19to110.java
@@ -861,8 +861,4 @@ public class MigrateFrom19to110 extends MigrationClientBase implements Migration
 
     }
 
-    @Override
-    public void preMigrationValidation(String validateStep) throws APIMigrationException {
-
-    }
 }
