+++ components/migration-client/wso2-api-migration-client/src/main/java/org/wso2/carbon/apimgt/migration/client/MigrateFrom200.java
@@ -108,11 +108,6 @@ public class MigrateFrom200 extends MigrationClientBase implements MigrationClie
 
     }
 
-    @Override
-    public void preMigrationValidation(String validateStep) throws APIMigrationException {
-
-    }
-
     public void addDefaultRoleCreationConfig() throws APIMigrationException {
         log.info("Add config in tenant-conf.json to enable default roles creation.");
         for (Tenant tenant : getTenantsArray()) {
