+++ components/migration-client/wso2-api-migration-client/src/main/java/org/wso2/carbon/apimgt/migration/client/MigrateFrom200to210.java
@@ -103,11 +103,6 @@ public class MigrateFrom200to210 extends MigrationClientBase implements Migratio
 
     }
 
-    @Override
-    public void preMigrationValidation(String validateStep) throws APIMigrationException {
-
-    }
-
     private void migrateFaultSequencesInRegistry() {
 
         /* change the APIMgtFaultHandler class name in debug_json_fault.xml and json_fault.xml
