+++ components/migration-client/wso2-api-migration-client/src/main/java/org/wso2/carbon/apimgt/migration/client/MigrateFrom210.java
@@ -106,8 +106,4 @@ public class MigrateFrom210 extends MigrationClientBase implements MigrationClie
     public void migrateTenantConfToDB() throws APIMigrationException {
 
     }
-
-    @Override public void preMigrationValidation(String validateStep) throws APIMigrationException {
-
-    }
 }
