+++ components/migration-client/wso2-api-migration-client/src/main/java/org/wso2/carbon/apimgt/migration/client/MigrateFrom310.java
@@ -124,11 +124,6 @@ public class MigrateFrom310 extends MigrationClientBase implements MigrationClie
     }
 
     @Override
-    public void preMigrationValidation(String validateStep) throws APIMigrationException {
-
-    }
-
-    @Override
     public void scopeMigration() throws APIMigrationException {
 
         APIMgtDAO apiMgtDAO = APIMgtDAO.getInstance();
