+++ components/migration-client/wso2-api-migration-client/src/main/java/org/wso2/carbon/apimgt/migration/client/MigrateFrom320.java
@@ -226,11 +226,6 @@ public class MigrateFrom320 extends MigrationClientBase implements MigrationClie
     }
 
     @Override
-    public void preMigrationValidation(String validateStep) throws APIMigrationException {
-        log.info("..........Pre migration steps are not defined for 3.2.0..........");
-    }
-
-    @Override
     public void scopeMigration() throws APIMigrationException {
     }
