+++ components/migration-client/wso2-api-migration-client/src/main/java/org/wso2/carbon/apimgt/migration/client/MigrationClient.java
@@ -122,5 +122,4 @@ public interface MigrationClient {
 
     void migrateTenantConfToDB() throws APIMigrationException;
 
-    void preMigrationValidation(String validateStep) throws APIMigrationException;
 }
