+++ components/migration-client/wso2-api-migration-client/src/main/java/org/wso2/carbon/apimgt/migration/client/ScopeRoleMappingPopulationClient.java
@@ -178,11 +178,6 @@ public class ScopeRoleMappingPopulationClient extends MigrationClientBase implem
     }
 
     @Override
-    public void preMigrationValidation(String validateStep) throws APIMigrationException {
-
-    }
-
-    @Override
     public void scopeMigration() throws APIMigrationException {
     }
