+++ components/migration-client/wso2-api-migration-client/src/main/java/org/wso2/carbon/apimgt/migration/client/sp_migration/APIMStatMigrationClient.java
@@ -200,9 +200,4 @@ public class APIMStatMigrationClient extends MigrationClientBase implements Migr
     public void migrateTenantConfToDB() throws APIMigrationException {
 
     }
-
-    @Override
-    public void preMigrationValidation(String validateStep) throws APIMigrationException {
-
-    }
 }
